Remove MinimiserBase<T>
